Extract HandlerThread implementation from Dispatcher
